+++ src/test/java/study/proxy/MethodMatcher.java
@@ -1,0 +1,9 @@
+package study.proxy;
+
+import java.lang.reflect.Method;
+
+@FunctionalInterface
+public interface MethodMatcher {
+
+    boolean matches(Method method, Class<?> targetClass, Object[] arguments);
+}
