+++ src/test/java/study/proxy/SayMethodMatcher.java
@@ -2,8 +2,16 @@ package study.proxy;
 
 import java.lang.reflect.Method;
 
-@FunctionalInterface
-public interface SayMethodMatcher {
+public class SayMethodMatcher implements MethodMatcher {
 
-    boolean matches(Method method, Class<?> targetClass, Object[] arguments);
+    private static final String PROXY_METHOD_PREFIX = "say";
+
+    @Override
+    public boolean matches(final Method method, final Class<?> targetClass, final Object[] arguments) {
+        return isMatch(method.getName());
+    }
+
+    private boolean isMatch(final String methodName) {
+        return methodName.startsWith(PROXY_METHOD_PREFIX);
+    }
 }
