+++ src/test/java/study/proxy/UppercaseInterceptor.java
@@ -7,23 +7,23 @@ import net.sf.cglib.proxy.MethodProxy;
 public class UppercaseInterceptor implements MethodInterceptor {
 
     private final Object target;
-    private final SayMethodMatcher sayMethodMatcher;
+    private final MethodMatcher methodMatcher;
 
     public UppercaseInterceptor(final Object target) {
         this.target = target;
-        this.sayMethodMatcher = new DefaultSayMethodMatcher();
+        this.methodMatcher = new SayMethodMatcher();
     }
 
-    public UppercaseInterceptor(final Object target, final SayMethodMatcher sayMethodMatcher) {
+    public UppercaseInterceptor(final Object target, final MethodMatcher methodMatcher) {
         this.target = target;
-        this.sayMethodMatcher = sayMethodMatcher;
+        this.methodMatcher = methodMatcher;
     }
 
     @Override
     public Object intercept(final Object obj, final Method method, final Object[] args, final MethodProxy proxy) throws Throwable {
         final Object result = proxy.invoke(target, args);
 
-        if (sayMethodMatcher.matches(method, target.getClass(), args)) {
+        if (methodMatcher.matches(method, target.getClass(), args)) {
             return ((String) result).toUpperCase();
         }
