+++ src/test/java/study/proxy/UppercaseInvocationHandler.java
@@ -6,23 +6,23 @@ import java.lang.reflect.Method;
 public class UppercaseInvocationHandler implements InvocationHandler {
 
     private final Object target;
-    private final SayMethodMatcher sayMethodMatcher;
+    private final MethodMatcher methodMatcher;
 
     public UppercaseInvocationHandler(final Object target) {
         this.target = target;
-        this.sayMethodMatcher = new DefaultSayMethodMatcher();
+        this.methodMatcher = new SayMethodMatcher();
     }
 
-    public UppercaseInvocationHandler(final Object target, final SayMethodMatcher sayMethodMatcher) {
+    public UppercaseInvocationHandler(final Object target, final MethodMatcher methodMatcher) {
         this.target = target;
-        this.sayMethodMatcher = sayMethodMatcher;
+        this.methodMatcher = methodMatcher;
     }
 
     @Override
     public Object invoke(final Object proxy, final Method method, final Object[] args) throws Throwable {
         final Object result = method.invoke(target, args);
 
-        if (sayMethodMatcher.matches(method, target.getClass(), args)) {
+        if (methodMatcher.matches(method, target.getClass(), args)) {
             return ((String) result).toUpperCase();
         }
